+++ src/__tests__/ZBClient.spec.ts
@@ -1,39 +1,43 @@
-import { ZBClient } from '..'
-import { clientOptions } from '../__mocks__/zbClientOptions'
+import { Loglevel, ZBClient } from '..'
+const clientOptions = {
+	loglevel: 'NONE' as Loglevel,
+}
+process.env.ZEEBE_NODE_LOG_LEVEL = 'NONE'
+let previousLogLevelEnv
 
 describe('ZBClient constructor', () => {
 	beforeEach(() => {
-		process.env.ZB_NODE_LOG_LEVEL = process.env.ZB_NODE_LOG_LEVEL || 'NONE'
+		process.env.ZEEBE_NODE_LOG_LEVEL =
+			process.env.ZEEBE_NODE_LOG_LEVEL || 'NONE'
+	})
+	beforeAll(() => {
+		previousLogLevelEnv = process.env.ZEEBE_NODE_LOG_LEVEL
+	})
+	afterAll(() => {
+		process.env.ZEEBE_NODE_LOG_LEVEL = previousLogLevelEnv
 	})
 	it('creates a new ZBClient', () => {
-		const zbc = new ZBClient('localhost')
+		const zbc = new ZBClient()
 		expect(zbc instanceof ZBClient).toBe(true)
 	})
 	it('appends the port number 26500 to the gatewayAddress by default', () => {
-		const zbc = new ZBClient('localhost')
+		const zbc = new ZBClient()
 		expect(zbc.gatewayAddress).toBe('localhost:26500')
 	})
 	it('accepts a custom port number for the gatewayAddress', () => {
 		const zbc = new ZBClient('localhost:123')
 		expect(zbc.gatewayAddress).toBe('localhost:123')
 	})
-	it('throws an exception when not provided a gatewayAddress in the constructor', () => {
-		expect(() => new (ZBClient as any)()()).toThrow()
-	})
-	it('takes client options passed in Ctor when ZB_NODE_LOG_LEVEL is not defined', () => {
-		process.env.ZB_NODE_LOG_LEVEL = ''
+	it('takes client options passed in Ctor when ZEEBE_NODE_LOG_LEVEL is not defined', () => {
+		process.env.ZEEBE_NODE_LOG_LEVEL = ''
 		clientOptions.loglevel = 'DEBUG'
-		const spy = jest.spyOn(clientOptions, 'loglevel', 'get')
-		expect(new ZBClient('localhost', clientOptions)).toBeInstanceOf(
-			ZBClient
-		)
-		expect(spy).toHaveBeenCalled()
-		spy.mockRestore()
+		const z = new ZBClient(clientOptions)
+		expect(z.loglevel).toBe('DEBUG')
 	})
-	it('ZB_NODE_LOG_LEVEL precedes options passed in Ctor', () => {
-		process.env.ZB_NODE_LOG_LEVEL = 'INFO'
+	it('ZEEBE_NODE_LOG_LEVEL precedes options passed in Ctor', () => {
+		process.env.ZEEBE_NODE_LOG_LEVEL = 'INFO'
 		clientOptions.loglevel = 'DEBUG'
-		const z = new ZBClient('localhost', clientOptions)
+		const z = new ZBClient(clientOptions)
 		expect(z.loglevel).toBe('INFO')
 	})
 })