+++ src/__tests__/integration/Client-DeployWorkflow.spec.ts
@@ -1,5 +1,6 @@
 import { ZBClient } from '../../index'
 const gatewayAddress = process.env.ZEEBE_GATEWAY_ADDRESS || '0.0.0.0:26500'
+process.env.ZEEBE_NODE_LOG_LEVEL = process.env.ZEEBE_NODE_LOG_LEVEL || 'NONE'
 
 describe('ZBClient', () => {
 	it('deploys a workflow', async () => {