+++ src/__tests__/integration/Client-MessageStart.spec.ts
@@ -1,7 +1,7 @@
 import { v4 as uuid } from 'uuid'
 import { ZBClient } from '../..'
 
-process.env.ZB_NODE_LOG_LEVEL = process.env.ZB_NODE_LOG_LEVEL || 'NONE'
+process.env.ZEEBE_NODE_LOG_LEVEL = process.env.ZEEBE_NODE_LOG_LEVEL || 'NONE'
 const gatewayAddress = process.env.ZEEBE_GATEWAY_ADDRESS || '0.0.0.0:26500'
 
 describe('ZBClient', () => {