+++ src/__tests__/integration/Worker-RaiseIncident.spec.ts
@@ -1,5 +1,6 @@
 import { ZBClient } from '../..'
 const gatewayAddress = process.env.ZEEBE_GATEWAY_ADDRESS || '0.0.0.0:26500'
+process.env.ZEEBE_NODE_LOG_LEVEL = process.env.ZEEBE_NODE_LOG_LEVEL || 'NONE'
 
 /**
  * Note: This test leaves its workflow instance active so the incident can be manually verified
@@ -55,7 +56,7 @@ describe('ZBWorker', () => {
 				// Manually verify that an incident has been raised
 				done()
 			},
-			{ pollInterval: 10000, maxJobsToActivate: 1, loglevel: 'NONE' }
+			{ longPoll: 10000, maxJobsToActivate: 1, loglevel: 'NONE' }
 		)
 	})
 })