+++ src/lib/GRPCClient.ts
@@ -10,9 +10,27 @@ import { ZBLogger } from './ZBLogger'
 interface GRPCClientExtendedOptions {
 	longPoll?: number
 }
+// tslint:disable: object-literal-sort-keys
 
-// @TODO: Better handling of status codes to deal with TLS and OAuth
-// https://github.com/grpc/grpc/blob/master/doc/statuscodes.md
+const GrpcError = {
+	OK: 0 as 0,
+	CANCELLED: 1 as 1,
+	UNKNOWN: 2 as 2,
+	INVALID_ARGUMENT: 3 as 3,
+	DEADLINE_EXCEEDED: 4 as 4,
+	NOT_FOUND: 5 as 5,
+	ALREADY_EXISTS: 6 as 6,
+	PERMISSION_DENIED: 7 as 7,
+	UNAUTHENTICATED: 16 as 16,
+	RESOURCE_EXHAUSTED: 8 as 8,
+	FAILED_PRECONDITION: 9 as 9,
+	ABORTED: 10 as 10,
+	OUT_OF_RANGE: 11 as 11,
+	UNIMPLEMENTED: 12 as 12,
+	INTERNAL: 13 as 13,
+	UNAVAILABLE: 14 as 14,
+	DATA_LOSS: 15 as 15,
+}
 
 const GrpcState = {
 	/**
@@ -49,14 +67,21 @@ const connectivityState = [
 export class GRPCClient extends EventEmitter {
 	public channelClosed = false
 	public longPoll?: number
+	public connected: boolean = false
 	public client: Client
+	public onReady?: () => void
 	private packageDefinition: PackageDefinition
 	private listNameMethods: string[]
 	private logger: ZBLogger
 	private gRPCRetryCount = 0
 	private oAuth?: OAuthProvider
+	private readyTimer?: NodeJS.Timeout
+	private failTimer?: NodeJS.Timeout
+	private connectionTolerance: number
+	private onConnectionError?: () => void
 
 	constructor({
+		connectionTolerance,
 		host,
 		loglevel,
 		oAuth,
@@ -66,7 +91,10 @@ export class GRPCClient extends EventEmitter {
 		service,
 		useTLS,
 		stdout = console,
+		onConnectionError,
+		onReady,
 	}: {
+		connectionTolerance: number
 		host: string
 		loglevel: Loglevel
 		oAuth?: OAuthProvider
@@ -76,10 +104,18 @@ export class GRPCClient extends EventEmitter {
 		service: string
 		useTLS: boolean
 		stdout: any
+		onConnectionError?: () => void
+		onReady?: () => void
 	}) {
 		super()
 		this.oAuth = oAuth
 		this.longPoll = options.longPoll
+		this.connectionTolerance = connectionTolerance
+
+		this.onReady = onReady
+		this.onConnectionError = onConnectionError
+		this.on('ready', () => this.setReady())
+		this.on('error', () => this.setNotReady())
 
 		this.logger = new ZBLogger({
 			color: chalk.green,
@@ -116,11 +152,6 @@ export class GRPCClient extends EventEmitter {
 				const methodName = listMethods[key].originalName as string
 
 				this.listNameMethods.push(methodName)
-
-				this[`${methodName}Async`] = async (data, fnAnswer) => {
-					const metadata = await this.getJWT()
-					this.client[methodName](data, metadata, fnAnswer)
-				}
 
 				this[`${methodName}Stream`] = async data => {
 					let stream
@@ -159,9 +190,20 @@ export class GRPCClient extends EventEmitter {
 						try {
 							const metadata = await this.getJWT()
 							client[methodName](data, metadata, (err, dat) => {
+								// This will error on network or business errors
 								if (err) {
+									const code = err.code
+									this.logger.error('Error: ', code)
+									const isNetworkError =
+										code === GrpcError.UNAVAILABLE
+									if (isNetworkError) {
+										this.setNotReady()
+									} else {
+										this.setReady()
+									}
 									return reject(err)
 								}
+								this.setReady()
 								resolve(dat)
 							})
 						} catch (e) {
@@ -242,12 +284,46 @@ export class GRPCClient extends EventEmitter {
 		})
 	}
 
+	private setReady() {
+		// debounce rapid connect / disconnect
+		if (this.readyTimer) {
+			clearTimeout(this.readyTimer)
+		}
+
+		this.readyTimer = setTimeout(() => {
+			this.readyTimer = undefined
+			this.connected = true
+			if (this.onReady) {
+				this.onReady()
+			}
+			if (this.failTimer) {
+				clearTimeout(this.failTimer)
+				this.failTimer = undefined
+			}
+		}, this.connectionTolerance)
+	}
+
+	private setNotReady() {
+		if (this.readyTimer) {
+			clearTimeout(this.readyTimer)
+			this.readyTimer = undefined
+		}
+		this.connected = false
+		if (!this.failTimer) {
+			this.failTimer = setTimeout(() => {
+				if (this.onConnectionError) {
+					this.onConnectionError()
+				}
+			}, this.connectionTolerance)
+		}
+	}
+
 	private handleGrpcError = (stream: any) => async (err: any) => {
 		this.emit('error', err)
 		this.logger.error(`GRPC ERROR: ${err.message}`)
 		const channelState = await this.watchGrpcChannel()
 		this.logger.debug(
-			`GRPC  Channel state: ${connectivityState[channelState]}`
+			`gRPC Channel state: ${connectivityState[channelState]}`
 		)
 		stream.removeAllListeners()
 		if (